RF_04001.091 | Move tests form Folder3Test to FolderTest
